+++ R/az_template.R
@@ -28,8 +28,10 @@
 #' - `parameters`: The parameters for the template. This can be provided using any of the same methods as the `template` argument.
 #' - `wait`: Optionally, whether to wait until the deployment is complete. Defaults to FALSE, in which case the method will return immediately.
 #'
+#' You can use the `build_template_definition` and `build_template_parameters` helper functions to construct the inputs for deploying a template. These can take as inputs R lists, JSON text strings, or file connections, and can also be extended by other packages.
+#'
 #' @seealso
-#' [az_resource_group], [az_resource],
+#' [az_resource_group], [az_resource], [build_template_definition], [build_template_parameters]
 #' [Template overview](https://docs.microsoft.com/en-us/azure/templates/),
 #' [Template API reference](https://docs.microsoft.com/en-us/rest/api/resources/deployments)
 #'
@@ -193,30 +195,41 @@ private=list(
         properties <- modifyList(default_properties, list(...))
         private$validate_deploy_parms(properties)
 
+        # rather than working with R objects, convert to JSON and do text munging
+        # this allows adding template/params that are already JSON text without conversion roundtrip
+        properties <- generate_json(properties)
+
+        # fold template data into properties
+        properties <- if(is.list(template))
+            append_json(properties, template=generate_json(template))
+        else if(is_file_spec(template))
+            append_json(properties, template=readLines(template))
+        else if(is_url(template))
+            append_json(properties, templateLink=generate_json(list(uri=template)))
+        else append_json(properties, template=template)
+
         # handle case of missing or empty parameters arg
         # must be a _named_ list for jsonlite to turn into an object, not an array
         if(missing(parameters) || is_empty(parameters))
-            parameters <- structure(list(), names=character(0))
+            parameters <- named_list()
 
-        # fold template data into list of properties
-        properties <- if(is.list(template))
-            modifyList(properties, list(template=template))
-        else if(is_url(template))
-            modifyList(properties, list(templateLink=list(uri=template)))
-        else modifyList(properties, list(template=jsonlite::fromJSON(template, simplifyVector=FALSE)))
-
-        # fold parameter data into list of properties
+        # fold parameter data into properties
         properties <- if(is_empty(parameters))
-            modifyList(properties, list(parameters=parameters))
+            append_json(properties, parameters=generate_json(parameters))
         else if(is.list(parameters))
-            modifyList(properties, list(parameters=private$make_param_list(parameters)))
+            append_json(properties, parameters=do.call(build_template_parameters, parameters))
+        else if(is_file_spec(parameters))
+            append_json(properties, parameters=readLines(parameters))
         else if(is_url(parameters))
-            modifyList(properties, list(parametersLink=list(uri=parameters)))
-        else modifyList(properties, list(parameters=jsonlite::fromJSON(parameters, simplifyVector=FALSE)))
+            append_json(properties, parametersLink=generate_json(list(uri=parameters)))
+        else append_json(properties, parameters=parameters)
 
         self$name <- name
-        parms <- private$tpl_op(body=jsonlite::toJSON(list(properties=properties), auto_unbox=TRUE, digits=22),
-            encode="raw", http_verb="PUT")
+        parms <- private$tpl_op(
+            body=jsonlite::prettify(sprintf('{"properties": %s}', properties)),
+            encode="raw",
+            http_verb="PUT"
+        )
 
         # do we wait until template has finished provisioning?
         if(wait)
@@ -270,12 +283,6 @@ private=list(
                     delete(confirm=FALSE, wait=TRUE))
             }
         }
-    },
-
-    # params for templates require lists of (value=x) rather than vectors as inputs
-    make_param_list=function(params)
-    {
-        lapply(params, function(x) if(is.list(x)) x else list(value=x))
     },
 
     tpl_op=function(op="", ...)